refactor: src/ui/ foundation folder — one file per reusable component
ui.rs → src/ui/{button,tab,badge,checkbox,overlay,menu,color,scrollbar}.rs.
Pulled the genuine primitives out of main.rs too (overlay, Badge + badge_inner
+ file_badge, checkbox_box), promoted to pub(crate), re-exported via ui/mod.rs
→ crate root. Non-UI helpers (pack_ext/pack_size) moved to modals.rs. Zero
behaviour change.
